requirements.txt
Django~=5.1.2
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
django-filter~=24.3
#Documentation
drf-yasg~=1.21.8